Added support for German, Spanish, & Italian localization(#21) in header labels.

+++ ReplyWithHeader/MHHeaderString.h
@@ -43,10 +43,8 @@
 
 - (id)initWithMailMessage:(id)mailMessage;
 - (void)applyHeaderTypography;
-//- (void)applyBoldFontTraits:(BOOL)isHeaderTypograbhyEnabled;
 - (void)applyHeaderLabelOptions;
 - (WebArchive *)getWebArchive;
-- (int)getHeaderItemCount;
-//- (BOOL)isSuppressLabelsFound;
+- (NSUInteger)getHeaderItemCount;
 
 @end
